week_06_07/project2_multimodel/requirements.txt
google-genai>=0.2.0   
transformers>=4.30.0
torch                
sentence-transformers
faiss-cpu
pytesseract
pillow
pdfplumber
opencv-python-headless
numpy
matplotlib
